+++ source/utils/types.js
@@ -4,3 +4,8 @@ import type { ElementRef } from 'react';
 export type ReactElementRef<ElementType = HTMLElement> = {
   current: null | ElementRef<ElementType>
 }
+
+export interface InputEvent {
+  target: HTMLInputElement,
+  inputType: string,
+}